Logistic and penalized logistic regression
I upload them but I continue check them. The problems with this
implementations is that the factor 1/N that we have to multiply the
gradient (as professor asked) causes problems. More specifically, the
algorithms have worse results than if we do not put this factor ( 75%
accuracy instead 99%, respectively)..... You can check it with the test
code that they gave us.