Removing outliers for regression task + Question
Here is the script to remove the outliers for the regression task. With
some simple changes we can perform the same for the classification.
Regarding the question that I have: since we have two models fitting in
our regression task, how we will compute the test error that he asks in
a .csv file?? If the error is the prediction of the error that the
unknown test data will have, my opinion is find the mean error per
sample for every model and after counting the samples that belong to
each of the models for the test data, we just multiply with the
respective portion factors  and we sum. What do you think??