Removing outliers - 2nd version
There is also this version of removing, where outliers are considered
more strictly. Here we exclude 0.7% of our total data, instead of 4.6%
in the other case. It seems (if I have implemented the predictions
right) that this removes some outliers at the X_test, too. You can
propose what we should keep.....